DelegatedComposer = Hanja + Emoji Composer
